Update parser generate states
After consuming a value we must check whether the next value is terminal
or non-termial. If it is non-terminal then we must add derived rules of
the non-terminal value to the items of the next state. In order to avoid
duplicate states we cannnot recurse into generate states until we have
finished building the state.