Add elided type to astnode struct
Abstract syntax tree can be less verbose than the full parse tree.  At
parse we still need to full parse type, however if we are reducing then
there is opportunity to elide types together for the abstract syntax
tree. Create an additional field in astnode struct to keep track of the
elided type.