Pass function declaration list through generator functions
In order to support local variables on the stack, other generator
functions need to know about where the variable was declared in the
function and saved on the stack.

Also be more strict about passing function arguments. These do not need
to be stored on the stack when there are less than 6 arguments. Instead
remember which register it is in. Registers are pushed/popped for nested
function calls as necessary.