Remove hardcoded stack local_size
Purpose of this commit is to get ready to support dynamic array
allocation. The overhead of keeping track of the size of the locally
allocated stack is unnecessary and will be a heavier burden when stack
size is dynamic.

NOTE: Seems this commit has known issue where following program doesn't
produce expected output...

    ```
    int main() {
        int i;
        int array[3];

        for (i=0; i<3; i++)
        {
            array[i] = i;
        }
        for (i=0; i<3; i++)
        {
            printf("%d\n", array[i]);
        }
    }
    ```