Update identifier_offset to return offset in register rbx
Purpose of this commit is to allow identifier_offset to be dynamically
calculated at runtime. We do this by returning the offset in rbx register. This
is the next step to allowing dynamic array allocation.

NOTE: This commit has known issue where following program doesn't
produce expected output...

    ```
    int main() {
        int *iptr;
        int i;
        i = 42;
        iptr = &i;
        printf("%d\n", *iptr);
        return 0;
    }
    ```

Issue is that local stack does not allocate enough space and so subsequent
register pushes onto the stack overwrite the local stack variables. Following
diff can fix this not adding additional (unallocated) offset to local
variables, but it causes hang in primes.c testcase.

    ```
    diff --git a/src/generator.c b/src/generator.c
    @@ -555,10 +555,6 @@ identifier_offset(struct ast_expression *ast,
             {
                 break;
             }
    -        /*
    -         * FIXME: Why is this necessary?
    -         */
    -        offset += 8;
         }

         write_assembly("  movq %%rbp, %%rbx");
    ```